requirements.txt
notion-client>=2.0.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
httpx>=0.25.0
watchdog>=3.0.0
Pygments>=2.16.0  # For code syntax highlighting
spotipy==2.23.0
